#49 : add tempPath to config

+++ slave/src/main/scala/slave/ResourceChecker.scala
@@ -9,16 +9,16 @@ class ResourceChecker {
   val runtime = Runtime.getRuntime
 
   // returns the memory that this program is using
-  def usedMemory:Int = (runtime.totalMemory - runtime.freeMemory()).toInt
+  def usedMemory:Long = (runtime.totalMemory - runtime.freeMemory())
  //??? can i use int ? or Long type.
 
   // returns the remaining memory that this program can use
-  def remainingMemory:Int = runtime.freeMemory.toInt
+  def remainingMemory:Long = runtime.freeMemory
 
   // returns the total memory that this program can use
-  def totalMemory:Int = runtime.totalMemory().toInt
+  def totalMemory:Long = runtime.totalMemory()
 
   // returns the Max Memory
-  def maxMemory : Int = runtime.maxMemory().toInt
+  def maxMemory : Long = runtime.maxMemory()
 }
